@@ -1,9 +1,10 @@
 import React from 'react';
+import bg from './images/eat-it-high-resolution-logo-color-on-transparent-background.png'
 
 function Header() {
     return (
         <div className="header">
-            <img src='../images/eat-it-high-resolution-logo-color-on-transparent-background.png'/>
+            <img src={bg}/>
         </div>
     )
 }
